Update to form image background and add Mysql.data and added DBUtilities class
I added the iphone background image to the form1 and change the property's for form 1 so the image fit the form. i also add the Mysql.Data package to the project in order to you my sql to pull and push data. Last thing that was add to the project was the DBUitiles class, the class will allow me to read the select server from a text file to connect to the data base.